修正 puzzle_new 函數簽名，重命名 puzzle_action 和 puzzle_is_win 函數以提高一致性，並更新 puzzle_snapshot 函數的返回類型

+++ src/lib.rs
@@ -25,7 +25,7 @@ pub extern "C" fn puzzle_is_win(puz: *mut Puzzle16) -> bool {
 }
 
 #[unsafe(no_mangle)]
-pub extern "C" fn puzzle_snapshot(puz: *const Puzzle16, out: *mut i32) -> usize {
+pub extern "C" fn puzzle_snapshot(puz: *const Puzzle16, out: *mut i32) -> u32 {
     use core::ptr::copy_nonoverlapping;
     if puz.is_null() || out.is_null() {
         return 0;
